+++ ssv/runner.go
@@ -24,7 +24,7 @@ type Runner interface {
 	Getters
 
 	// StartNewDuty starts a new duty for the runner, returns error if can't
-	StartNewDuty(duty types.Duty, quorum int) error
+	StartNewDuty(duty types.Duty, quorum uint64) error
 	// HasRunningDuty returns true if it has a running duty
 	HasRunningDuty() bool
 	// ProcessPreConsensus processes all pre-consensus msgs, returns error if can't process
@@ -79,12 +79,12 @@ func (b *BaseRunner) SetHighestDecidedSlot(slot spec.Slot) {
 }
 
 // setupForNewDuty is sets the runner for a new duty
-func (b *BaseRunner) baseSetupForNewDuty(duty types.Duty, quorum int) {
+func (b *BaseRunner) baseSetupForNewDuty(duty types.Duty, quorum uint64) {
 	b.State = NewRunnerState(quorum, duty)
 }
 
 // baseStartNewDuty is a base func that all runner implementation can call to start a duty
-func (b *BaseRunner) baseStartNewDuty(runner Runner, duty types.Duty, quorum int) error {
+func (b *BaseRunner) baseStartNewDuty(runner Runner, duty types.Duty, quorum uint64) error {
 	if err := b.ShouldProcessDuty(duty); err != nil {
 		return errors.Wrap(err, "can't start duty")
 	}
@@ -95,7 +95,7 @@ func (b *BaseRunner) baseStartNewDuty(runner Runner, duty types.Duty, quorum int
 }
 
 // baseStartNewBeaconDuty is a base func that all runner implementation can call to start a non-beacon duty
-func (b *BaseRunner) baseStartNewNonBeaconDuty(runner Runner, duty *types.BeaconDuty, quorum int) error {
+func (b *BaseRunner) baseStartNewNonBeaconDuty(runner Runner, duty *types.BeaconDuty, quorum uint64) error {
 	if err := b.ShouldProcessNonBeaconDuty(duty); err != nil {
 		return errors.Wrap(err, "can't start non-beacon duty")
 	}
